Only reset election timeout if we grant vote
bc, for Followers, the paper says

"If election timeout elapses without receiving AppendEntries
RPC from current leader or granting vote to candidate:
convert to candidate"

This is implying we shouldn't reset the timeout just becuase we got a vote request. We should definitely reset it if we granted the vote though.

Tests are still passing